Separated out diffusion training functionality into rtk.diffusion. Still trying to fix metrics, but it seems that the ordering of the fake_images is wrong as it's [128,128,3] when it should be [3,128,128]
