+++ configs/tests.yaml
@@ -4,7 +4,7 @@ defaults:
   - _self_
   - datasets: chest-xray14-dataset
   - job: eval-diffusion
-  - models: default-model
+  - models: ddpm-model
   # module configurations
   - ignite: ignite
   - mlflow: local
@@ -24,6 +24,7 @@ hydra:
     chdir: true
   searchpath:
     - file:///home/nicoleg/workspaces/dissertation/configs
+    - file:///home/nicoleg/workspaces/rcg/config
   sweep:
     dir: outputs/${hydra.job.config_name}/${date}/${timestamp}
     subdir: ${hydra.job.override_dirname}
